\name{albumin}
\alias{albumin}
\docType{data}
\title{
Albumin data.  
}
\description{
Data from four experiments about the percentage of albumin in the plasma protein of the normal human subjects.
}
\usage{data(albumin)}
\format{
\describe{
\item{\code{y}}{mean albumin percentage.}
\item{\code{sigma2}}{estimated within-study variance.}
}
}
\source{
Meier, P. (1953). Variance of a weighted mean. \emph{Biometrics} \bold{9}, 59--73.
}
\examples{
data(albumin)
}
\keyword{datasets}