\name{diuretics}
\alias{diuretics}
\docType{data}
\title{
Diuretics data.  
}
\description{
Data from nine randomized trials on prevention of pre-eclampsia with diuretics.   
}
\usage{data(diuretics)}
\format{
\describe{
\item{\code{y}}{logarithm of the risk ratio in each study.}
\item{\code{sigma2}}{estimated within-study variance.}
}
}
\source{
Biggerstaff, B. and Tweedie, R. (1997). Incorporating variability in estimates of heterogeneity in the random effects model in meta-analysis. \emph{Statistics in Medicine} \bold{16}, 753--768.
}
\examples{
data(diuretics)
}
\keyword{datasets}