\name{education}
\alias{education}
\docType{data}
\title{
Open education data.  
}
\description{
Data from eleven studies on the effect of open versus traditional education on student attitude toward schools.
}
\usage{data(education)}
\format{
\describe{
\item{\code{y}}{standardized estimated mean difference in attitude according to the type of education.}
\item{\code{sigma2}}{estimated within-study variance.}
}
}
\source{
Hedges, L.V. and Olkin, I. (1985). \emph{Statistical Methods for Meta-Analysis}. Academic Press, Orlando.
}
\examples{
data(education)
}
\keyword{datasets}